Routing implemented to the navbar options for  DevLogs, About, and the BTC/XMR donation pages.Defauth Home page can be seen by clicking on simple snake.  To mantain consistencey in Navbar UI with added Link Tags I wrote some css with class Name like in saperate navebar.css file
