+++ src/App.css
@@ -1,44 +1,46 @@
+/** @format */
+
 .app {
-  padding: 10px;
+	padding: 10px;
 }
 
 .main {
-  min-height: 100%;
+	min-height: 100%;
 }
 
 form {
-  width: 50%;
+	width: 50%;
 }
 
 .form-input {
-  margin: 10px 0px;
+	margin: 10px 0px;
 }
 
 .form-input input,
 .form-input textarea {
-  width: 100%;
-  font-size: 1rem;
-  line-height: 1.5;
-  border-radius: 0.25rem;
-  border: 1px solid lightgray;
-  padding: 0.375rem 0.75rem;
-  outline: none;
+	width: 100%;
+	font-size: 1rem;
+	line-height: 1.5;
+	border-radius: 0.25rem;
+	border: 1px solid lightgray;
+	padding: 0.375rem 0.75rem;
+	outline: none;
 }
 
 button {
-  background-color: #007bff;
-  color: #fff;
-  padding: 0.375rem 0.75rem;
-  font-size: 1rem;
-  line-height: 1.5;
-  border-radius: 0.25rem;
-  cursor: pointer;
-  outline: none;
+	background-color: #007bff;
+	color: #fff;
+	padding: 0.375rem 0.75rem;
+	font-size: 1rem;
+	line-height: 1.5;
+	border-radius: 0.25rem;
+	cursor: pointer;
+	outline: none;
 }
 
 .blog-post__display {
-  border: 1px solid lightgrey;
-  margin: 5px 0px;
-  border-radius: 4px;
-  padding: 10px;
+	border: 1px solid lightgrey;
+	margin: 5px 0px;
+	border-radius: 4px;
+	padding: 10px;
 }
