+++ src/components/Navebar.css
@@ -1,0 +1,10 @@
+/** @format */
+
+.link {
+	color: white;
+	text-decoration: none;
+}
+.link:hover {
+	color: wheat;
+	text-decoration: none;
+}
