+++ src/pages/About/About.js
@@ -1,0 +1,9 @@
+/** @format */
+
+import React from 'react';
+
+const About = () => {
+	return <h1>About Page</h1>;
+};
+
+export default About;
