+++ src/pages/BTC/Bitcoin.js
@@ -1,0 +1,9 @@
+/** @format */
+
+import React from 'react';
+
+const Bitcoin = () => {
+	return <h1>Bitcoin Page</h1>;
+};
+
+export default Bitcoin;
