+++ src/pages/XMR/xmr.js
@@ -1,0 +1,9 @@
+/** @format */
+
+import React from 'react';
+
+const Xmr = () => {
+	return <h1>Xmr Page</h1>;
+};
+
+export default Xmr;
